+++ Projects/ProjectARVRPro/Process/MTFHV058/MTFHV058RecipeConfig.cs
@@ -1,0 +1,87 @@
+
+using ColorVision.Common.MVVM;
+using ProjectARVRPro.Recipe;
+using System.ComponentModel;
+
+namespace ProjectARVRPro.Process.MTFHV058
+{
+    public class MTFHV058RecipeConfig : ViewModelBase, IRecipeConfig
+    {
+        [Category("MTF058")]
+        public RecipeBase MTF058_H_Center_0F { get => _MTF058_H_Center_0F; set { _MTF058_H_Center_0F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_H_Center_0F = new RecipeBase(0.5, 0);
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_V_Center_0F { get => _MTF058_V_Center_0F; set { _MTF058_V_Center_0F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_V_Center_0F = new RecipeBase(0.5, 0);
+
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_H_LeftUp_0_5F { get => _MTF_HV_H_LeftUp_0_5F; set { _MTF_HV_H_LeftUp_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_H_LeftUp_0_5F = new RecipeBase(0.5, 0);
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_V_LeftUp_0_5F { get => _MTF_HV_V_LeftUp_0_5F; set { _MTF_HV_V_LeftUp_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_V_LeftUp_0_5F = new RecipeBase(0.5, 0);
+
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_H_RightUp_0_5F { get => _MTF_HV_H_RightUp_0_5F; set { _MTF_HV_H_RightUp_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_H_RightUp_0_5F = new RecipeBase(0.5, 0);
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_V_RightUp_0_5F { get => _MTF_HV_V_RightUp_0_5F; set { _MTF_HV_V_RightUp_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_V_RightUp_0_5F = new RecipeBase(0.5, 0);
+
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_H_RightDown_0_5F { get => _MTF_HV_H_RightDown_0_5F; set { _MTF_HV_H_RightDown_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_H_RightDown_0_5F = new RecipeBase(0.5, 0);
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_V_RightDown_0_5F { get => _MTF_HV_V_RightDown_0_5F; set { _MTF_HV_V_RightDown_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_V_RightDown_0_5F = new RecipeBase(0.5, 0);
+
+        
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_H_LeftDown_0_5F { get => _MTF_HV_H_LeftDown_0_5F; set { _MTF_HV_H_LeftDown_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_H_LeftDown_0_5F = new RecipeBase(0.5, 0);
+
+        [Category("MTF_HV_058")]
+        public RecipeBase MTF_HV_V_LeftDown_0_5F { get => _MTF_HV_V_LeftDown_0_5F; set { _MTF_HV_V_LeftDown_0_5F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF_HV_V_LeftDown_0_5F = new RecipeBase(0.5, 0);
+
+
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_H_LeftUp_0_8F { get => _MTF058_H_LeftUp_0_8F; set { _MTF058_H_LeftUp_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_H_LeftUp_0_8F = new RecipeBase(0.5, 0);
+        [Category("MTF058")]
+        public RecipeBase MTF058_V_LeftUp_0_8F { get => _MTF058_V_LeftUp_0_8F; set { _MTF058_V_LeftUp_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_V_LeftUp_0_8F = new RecipeBase(0.5, 0);
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_H_RightUp_0_8F { get => _MTF058_H_RightUp_0_8F; set { _MTF058_H_RightUp_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_H_RightUp_0_8F = new RecipeBase(0.5, 0);
+
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_V_RightUp_0_8F { get => _MTF058_V_RightUp_0_8F; set { _MTF058_V_RightUp_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_V_RightUp_0_8F = new RecipeBase(0.5, 0);
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_H_RightDown_0_8F { get => _MTF058_H_RightDown_0_8F; set { _MTF058_H_RightDown_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_H_RightDown_0_8F = new RecipeBase(0.5, 0);
+        [Category("MTF058")]
+        public RecipeBase MTF058_V_RightDown_0_8F { get => _MTF058_V_RightDown_0_8F; set { _MTF058_V_RightDown_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_V_RightDown_0_8F = new RecipeBase(0.5, 0);
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_H_LeftDown_0_8F { get => _MTF058_H_LeftDown_0_8F; set { _MTF058_H_LeftDown_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_H_LeftDown_0_8F = new RecipeBase(0.5, 0);
+
+        [Category("MTF058")]
+        public RecipeBase MTF058_V_LeftDown_0_8F { get => _MTF058_V_LeftDown_0_8F; set { _MTF058_V_LeftDown_0_8F = value; OnPropertyChanged(); } }
+        private RecipeBase _MTF058_V_LeftDown_0_8F = new RecipeBase(0.5, 0);
+    }
+}